Create Temp Mock

+++ app/src/main/java/com/example/shiftproj/model/entity/CityTemp.kt
@@ -2,4 +2,4 @@ package com.example.shiftproj.model.entity
 
 import java.io.Serializable
 
-data class CityTemp (val city : String, val temp : Int = -1, val id : Int = -1) : Serializable
+data class CityTemp (val city : String, val temp : Double = -1.0, val id : Int = -1) : Serializable
