+++ app/src/main/java/com/example/shiftproj/presentation/ui/citesTemp/add/AddCityTempView.kt
@@ -3,5 +3,5 @@ package com.example.shiftproj.presentation.ui.citesTemp.add
 import com.example.shiftproj.presentation.ui.ViewContract
 
 interface AddCityTempView : ViewContract {
-    fun onAddButtonClick(city : String, temp : Int)
+    fun onAddButtonClick(city : String, temp : Double)
 }
